Added color strip and outline for property draw.
The property draw method now creates a colored strip at
the top. A black outline is also drawn around the property.

A bunch of constants for configuration of the drawing code are
going to be defined. Unfortunately this takes up code space
since there is no preprocessor in java that I know of,
but I am not worried about that small amount of memory ( a few bytes).

I am not sure what to do with the GameBuffer situation.
Should each property have an x and a y and all draw on one
big "Board GameBuffer", or should each property draw on their
own buffer that gets copied onto the Board's buffer somewhere?
I will investigate this further.