Removed Property. Everything will be a Space.
I decided to get rid of Property.
I felt that it proved to be a superfluous distinction.
Space will have a way to not draw anything but a custom
image for the corner spaces.
Additionally, I made an executive decision to have the board
be made entirely of squares, because that greatly simplifies the
math I have to do to draw it.